requirements.txt
# Core Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0

# AI and Machine Learning
openai>=1.3.0
langchain>=0.0.350
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.2
groq>=0.4.2  # Added for Groq LLM support

# Memory and Storage
chromadb>=0.4.22
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Utilities
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
tenacity>=8.2.3
loguru>=0.7.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.1

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0